a codegen invokation, for early experimentation.
the codegen tools don't yet have a CLI that makes them more usable;
so, instead, I invoke the code directly.
Doing this from a test function means running 'go test' on that package
will make things happen -- kludgey, but functional.
(Doing it in a sub-package means if there's compile errors in the
output, it doens't wedge regen; that package starting with an
underscore means the regen only happens if you dive specifically
into that package (it's not caught by a './...').)

I'll commit the results in the next commit.