initial commit implementing ipld.Node
This is the first working version of a HAMT ADL using go-ipld-prime. See
basic_test.go for some examples.

The main feature implemented is building an entirely new HAMT via the
builder/assembler interfaces. Loading existing HAMTs, or modifying
existing HAMTs, are not yet implemented.

Note that this code implements a HAMT following the schema described at
https://github.com/ipld/specs/blob/HEAD/data-structures/hashmap.md,
which is a different schema than Filecoin's HAMT.