first working version of filecoin/go-hamt-ipld compatibility
If FilecoinV3Prototype is used instead of Prototype, then we build a
Filecoin-compatible HAMT with hard-coded parameters at build time. The
parameters are omitted from the block encoding too, just like
Filecoin's. Finally, the bitmap is used via a *big.Int.

The test only includes building a simple map. For the compatibility with
Filecoin HAMTs to be useful, a few pieces of follow-up work would be
necessary:

1) Reification, to load a Filecoin HAMT in its block format and use our
   ADL implementation to query the map.

2) More tests, particularly loading complex and multi-block HAMTs
   created by go-hamt-ipld.

3) Refinement of the internals. For example, sorting of buckets and
   caching of links are TODOs.